requirements.txt
aiogram[aiohttp]~=3.4
aiohttp
requests
python-dotenv